+++ front/app/community/[siren]/components/FicheSubventions/RankingTable.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { Subvention } from '#app/models/subvention';
-import { PaginationProps, WithPagination } from '#components/Pagination';
+import { WithPagination } from '#components/Pagination';
 import { Badge } from '#components/ui/badge';
 import {
   Table as ShadCNTable,
@@ -11,92 +10,96 @@ import {
   TableHeader,
   TableRow,
 } from '#components/ui/table';
+import { usePagination } from '#utils/hooks/usePagination';
+import { useSubventionPaginated } from '#utils/hooks/useSubventionPaginated';
 import { formatAmount } from '#utils/utils';
 
 import { YearOption } from '../../types/interface';
-import { CHART_HEIGHT } from '../constants';
 import { NoData } from '../NoData';
+import MarchesPublicsTableSkeleton from '../Skeletons/MarchesPublicsTableSkeleton';
+import { CHART_HEIGHT } from '../constants';
 
-const ROWS_COUNT = 10;
-
-interface RankingTableProps {
-  data: Subvention[];
+type SubventionTableProps = {
+  siren: string;
   year: YearOption;
-  paginationProps: Omit<PaginationProps, 'totalPage'>;
-}
+  paginationProps: ReturnType<typeof usePagination>;
+};
 
-export default function RankingTable({ data, year, paginationProps }: RankingTableProps) {
-  const filteredData =
-    year === 'All' ? data : data.filter((item) => String(item.annee) === String(year));
+const MAX_ROW_PER_PAGE = 10;
 
-  function getShownSubs(data: Subvention[]) {
-    const sortedSubs = data.sort((a, b) => Number(b.montant) - Number(a.montant));
-    const subs = sortedSubs.slice(
-      ROWS_COUNT * (paginationProps.activePage - 1),
-      ROWS_COUNT * paginationProps.activePage,
+export default function RankingTable({ siren, year, paginationProps }: SubventionTableProps) {
+  const { data, isPending, isError } = useSubventionPaginated(siren, year === 'All' ? null : year, {
+    page: paginationProps.activePage,
+    limit: MAX_ROW_PER_PAGE,
+  });
+
+  if (isPending || isError) {
+    return (
+      <div style={{ height: CHART_HEIGHT }}>
+        <MarchesPublicsTableSkeleton rows={MAX_ROW_PER_PAGE} />
+      </div>
     );
-
-    return subs;
   }
 
-  const totalPage = Math.ceil(filteredData.length / ROWS_COUNT);
-  const rows = getShownSubs(filteredData);
-
-  if (rows.length === 0) {
+  if (data.length === 0) {
     return <NoData style={{ height: CHART_HEIGHT }} />;
   }
 
+  const rows: Row[] = data.map(
+    ({ id_beneficiaire, beneficiaire_names, objet, montant, annee }) => ({
+      id: id_beneficiaire,
+      names: beneficiaire_names,
+      object: objet,
+      amount: montant,
+      year: annee,
+    }),
+  );
+
+  const totalPage = Math.ceil(data[0].total_row_count / MAX_ROW_PER_PAGE);
+
   return (
-    <>
-      <WithPagination style={{ height: CHART_HEIGHT }} totalPage={totalPage} {...paginationProps}>
-        <Table rows={rows} />
-      </WithPagination>
-    </>
+    <WithPagination style={{ height: CHART_HEIGHT }} totalPage={totalPage} {...paginationProps}>
+      <Table rows={rows} />
+    </WithPagination>
   );
 }
 
-function Table({ rows }: { rows: Subvention[] }) {
-  function formatSubventionObject(input: string): string[] {
-    return input
-      .replace(/[\[\]]/g, '') // Supprime les crochets
-      .replace(/\\r\\n|\r\n|\n/g, ' ') // Retire les \n\r
-      .split(/',|",/) // Split sur des virgules
-      .map((item) =>
-        item
-          .trim()
-          .replace(/^['"]|['"]$/g, '')
-          .toLocaleUpperCase(),
-      );
-  }
+type Row = {
+  id: string | number;
+  names: string[];
+  object: string;
+  amount: number;
+  year: number;
+};
 
+type Table = {
+  rows: Row[];
+};
+
+export function Table({ rows }: Table) {
   return (
     <ShadCNTable>
       <TableHeader>
         <TableRow>
-          <TableHead className='w-[300px]'>Bénéficiaires</TableHead>
+          <TableHead className='w-[300px]'>Bénéficaire</TableHead>
           <TableHead className=''>Objet</TableHead>
           <TableHead className='w-[140px] text-right'>Montant (€)</TableHead>
           <TableHead className='w-[140px] text-right'>Année</TableHead>
         </TableRow>
       </TableHeader>
       <TableBody>
-        {rows.map((item, index) => (
-          <TableRow key={index}>
-            <TableCell className='font-medium'>
-              { item.nom_beneficiaire ? <Badge className='m-1'>
-                {item.nom_beneficiaire}
-              </Badge> : null }
-            </TableCell>
+        {rows.map(({ id, names, object, amount, year }) => (
+          <TableRow key={`${id}-${object}-${year}`}>
             <TableCell>
-              {formatSubventionObject(item.objet).map((item, index) => (
-                <span key={index}>
-                  {index > 0 && ' - '}
-                  {item}
-                </span>
+              {names.map((name) => (
+                <Badge key={name} className='m-1'>
+                  {name}
+                </Badge>
               ))}
             </TableCell>
-            <TableCell className='text-right'>{formatAmount(item.montant)}</TableCell>
-            <TableCell className='text-right'>{item.annee}</TableCell>
+            <TableCell>{object.toLocaleUpperCase()}</TableCell>
+            <TableCell className='text-right'>{formatAmount(amount)}</TableCell>
+            <TableCell className='text-right'>{year}</TableCell>
           </TableRow>
         ))}
       </TableBody>
