+++ front/app/community/[siren]/components/FicheSubventions/SubventionsWithState.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { Subvention } from '#app/models/subvention';
 import { TransparencyScore } from '#components/TransparencyScore/constants';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '#components/ui/tabs';
 import { TAB_VALUES, useSubventionsTab } from '#hooks/useTabState';
@@ -18,34 +17,22 @@ interface SubventionsWithStateProps {
   communityType: CommunityType;
 }
 
-export function SubventionsWithState({ siren, subventions, availableYears, communityType }: SubventionsWithStateProps) {
+export function SubventionsWithState({ siren, availableYears, communityType }: SubventionsWithStateProps) {
   const [activeTab, setActiveTab] = useSubventionsTab();
 
   return (
     <Tabs value={activeTab} onValueChange={setActiveTab}>
-      <TabsList className='h-10 p-0.5 sm:h-12 sm:p-1'>
-        <TabsTrigger
-          value={TAB_VALUES.SUBVENTIONS.TRENDS}
-          className='px-2 text-xs sm:px-3 sm:text-sm'
-        >
+      <TabsList className='h-10 sm:h-12 p-0.5 sm:p-1'>
+        <TabsTrigger value={TAB_VALUES.SUBVENTIONS.TRENDS} className='px-2 sm:px-3 text-xs sm:text-sm'>
           Évolution
         </TabsTrigger>
-        <TabsTrigger
-          value={TAB_VALUES.SUBVENTIONS.DISTRIBUTION}
-          className='px-2 text-xs sm:px-3 sm:text-sm'
-        >
+        <TabsTrigger value={TAB_VALUES.SUBVENTIONS.DISTRIBUTION} className='px-2 sm:px-3 text-xs sm:text-sm'>
           Répartition
         </TabsTrigger>
-        <TabsTrigger
-          value={TAB_VALUES.SUBVENTIONS.COMPARISON}
-          className='px-2 text-xs sm:px-3 sm:text-sm'
-        >
+        <TabsTrigger value={TAB_VALUES.SUBVENTIONS.COMPARISON} className='px-2 sm:px-3 text-xs sm:text-sm'>
           Comparaison
         </TabsTrigger>
-        <TabsTrigger
-          value={TAB_VALUES.SUBVENTIONS.DETAILS}
-          className='px-2 text-xs sm:px-3 sm:text-sm'
-        >
+        <TabsTrigger value={TAB_VALUES.SUBVENTIONS.DETAILS} className='px-2 sm:px-3 text-xs sm:text-sm'>
           Classement
         </TabsTrigger>
       </TabsList>
